+++ lib/src/deprecation.dart
@@ -54,9 +54,9 @@ enum Deprecation {
       deprecatedIn: '1.56.0',
       description: 'Passing invalid units to built-in functions.'),
 
-  /// Deprecation for passing invalid units to certain built-in functions.
+  /// Deprecation for passing percentages to the Sass abs() function.
   absPercent('abs-percent',
-      deprecatedIn: '1.62.0',
+      deprecatedIn: '1.64.0',
       description: 'Passing percentages to the Sass abs() function.'),
 
   duplicateVariableFlags('duplicate-var-flags',